Wrapper to Extension, closes #1853
